Moved disjoint models to central folder
