+++ src/app/models/period.model.ts
