+++ src/app/models/schedule-event.model.ts
